@@ -1,17 +1,15 @@
 package pb.ajneb97.admin;
 
-import java.util.ArrayList;
 import org.bukkit.Bukkit;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.scheduler.BukkitScheduler;
 
 import pb.ajneb97.PaintballBattle;
-import pb.ajneb97.logic.TopHologram;
 
 public class TopHologramAdmin {
 
 	int taskID;
-	private PaintballBattle plugin;
+	private final PaintballBattle plugin;
 	public TopHologramAdmin(PaintballBattle plugin){		
 		this.plugin = plugin;		
 	}
@@ -20,22 +18,18 @@ public class TopHologramAdmin {
 		return this.taskID;
 	}
 	
-	public void updateHolograms() {
+	public void scheduledUpdateHolograms() {
 		FileConfiguration config = plugin.getConfig();
-		long ticks = Long.valueOf(config.getString("top_hologram_update_time"))*20;
+		long topHologramUpdateTimeInTicks = Long.parseLong(config.getString("top_hologram_update_time"))*20;
 	    BukkitScheduler scheduler = Bukkit.getServer().getScheduler();
  	    taskID = scheduler.scheduleSyncRepeatingTask(plugin, new Runnable() {
             public void run() { 
-            	ejecutarActualizarHologramas();
+            	updateHolograms();
             }
-        },ticks, ticks);
+        }, topHologramUpdateTimeInTicks, topHologramUpdateTimeInTicks);
 	}
 
-	protected void ejecutarActualizarHologramas() {
-		ArrayList<TopHologram> hologramas = plugin.getTopHolograms();
-		for(int i=0;i<hologramas.size();i++) {
-			hologramas.get(i).actualizar(plugin);
-			
-		}
+	protected void updateHolograms() {
+		plugin.getTopHolograms().forEach(topHologram -> topHologram.update(plugin));
 	}
 }
