@@ -41,15 +41,15 @@ public class SignInteractEventHandler implements Listener{
 						FileConfiguration messages = plugin.getMessages();
 						PaintballMatch paintballMatch = plugin.getMatch(arena);
 						String estado = "";
-						if(paintballMatch.getEstado().equals(MatchStatus.PLAYING)) {
+						if(paintballMatch.getState().equals(MatchStatus.PLAYING)) {
 							estado = messages.getString("signStatusIngame");
-						}else if(paintballMatch.getEstado().equals(MatchStatus.STARTING)) {
+						}else if(paintballMatch.getState().equals(MatchStatus.STARTING)) {
 							estado = messages.getString("signStatusStarting");
-						}else if(paintballMatch.getEstado().equals(MatchStatus.WAITING)) {
+						}else if(paintballMatch.getState().equals(MatchStatus.WAITING)) {
 							estado = messages.getString("signStatusWaiting");
-						}else if(paintballMatch.getEstado().equals(MatchStatus.OFF)) {
+						}else if(paintballMatch.getState().equals(MatchStatus.OFF)) {
 							estado = messages.getString("signStatusDisabled");
-						}else if(paintballMatch.getEstado().equals(MatchStatus.ENDING)) {
+						}else if(paintballMatch.getState().equals(MatchStatus.ENDING)) {
 							estado = messages.getString("signStatusFinishing");
 						}
 						
